backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
openai>=2.8.0
pandas==2.1.3
numpy==1.26.2
faiss-cpu==1.7.4
rank-bm25==0.2.2
sentence-transformers>=5.1.0
pydantic==2.5.0
python-multipart==0.0.6
httpx==0.25.1
exa-py==2.0.1
